All but one left
