+++ frontend/src/components/ui/button.variants.ts
@@ -1,0 +1,29 @@
+import { cva } from "class-variance-authority";
+
+export const buttonVariants = cva(
+  "inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium transition-all disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg:not([class*='size-'])]:size-4 shrink-0 [&_svg]:shrink-0 outline-none focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive",
+  {
+    variants: {
+      variant: {
+        default:
+          "bg-highlight2 text-highlight2 border border-transparent shadow-xs px-3 py-1 hover:text-highlight2 transition-colors",
+        destructive:
+          "bg-destructive text-white rounded-[1rem] shadow-xs px-3 py-1 border border-transparent hover:bg-destructive/90 focus-visible:ring-destructive/20 dark:focus-visible:ring-destructive/40 dark:bg-destructive/60",
+        outline:
+          "bg-highlight2 text-white rounded-[1rem] border border-transparent shadow-xs px-3 py-1 hover:bg-white hover:text-highlight2 hover:border-highlight2 transition-colors",
+        secondary:
+          "bg-white text-secondary rounded-[1rem] border border-secondary shadow-xs px-3 py-1 hover:bg-secondary hover:text-white hover:border-secondary",
+        ghost:
+          "bg-transparent text-secondary hover:bg-secondary hover:text-white px-3 py-1 transition-colors",
+      },
+      size: {
+        sm: "h-8 px-3 text-sm",
+        md: "h-10 px-3 text-sm",
+        lg: "h-12 px-6 text-base",
+      },
+    },
+    defaultVariants: {
+      size: "sm",
+    },
+  },
+);
